pubsub/requirements.txt
google-cloud-pubsub~=0.39.0